ai_service/requirements.txt

flask
flask-cors
ultralytics
tensorflow
opencv-python
numpy
pillow
